Break CLI into own file
